// froid-cheikh-ecommerce-frontend/src/app/models/interfaces.ts
export enum LabelEnergie {
  A_PLUS_PLUS_PLUS = 'A_PLUS_PLUS_PLUS',
  A_PLUS_PLUS = 'A_PLUS_PLUS',
  A_PLUS = 'A_PLUS',
  A = 'A',
  B = 'B',
  C = 'C',
  D = 'D',
  E = 'E',
  F = 'F',
  G = 'G'
}
// ...
export interface Produit {
  idProduit: number;
  nomProduit: string;
  descriptionProduit?: string;
  prix: number;
  stockDisponible: number;
  refProduit?: string;
  codeProduit?: string;
  garantie?: string;
  labelEnergie?: LabelEnergie;
  puissanceBTU?: number;
  consommationWatt?: number;
  dimensions?: string;
  poids?: number;
  ficheTechniquePDF?: string;
  listeImages: string[];
  videosOptionnelles?: string[];
  disponibilite: boolean;
  dateAjout: string;
  categorie?: Categorie;
  marque?: Marque;
  attributs?: AttributProduit[];
}

export interface AttributProduit {
  idAttribut: number;
  nomAttribut: string;
  valeurAttribut: string;
}

export interface Categorie {
  idCategorie: number;
  nomCategorie: string;
  descriptionCategorie?: string;
  imageCategorie?: string;
  parentId?: number;
  nomParent?: string;
  sousCategories?: Categorie[];
  nombreProduits?: number;
}

export interface Marque {
  idMarque: number;
  nomMarque: string;
  logo?: string;
  description?: string;
  nombreProduits?: number;
}